% Generated by roxygen2 (4.0.1): do not edit by hand
\name{copy_object}
\alias{copy_object}
\title{Copy an object from one bucket to another.}
\usage{
copy_object(sourceBucketName, sourceKey, destinationBucketName, destinationKey)
}
\arguments{
\item{sourceBucketName}{The name of the bucket the file is in originally.}

\item{sourceKey}{The location the file is including folder as well as the
file name.}

\item{destinationBucketName}{The name of the bucket you want the file copied to.}

\item{destinationKey}{The location you want the file copied too.}
}
\value{
Returns 1 if successful
}
\description{
Copy an object from one bucket to another.
}
\details{
This function copies a file from one bucket to another
bucket and can place it within a new folder.
}

