% Generated by roxygen2 (4.0.1): do not edit by hand
\name{delete_bucket}
\alias{delete_bucket}
\title{Delete a bucket that you have.}
\usage{
delete_bucket(bucketName)
}
\arguments{
\item{bucketName}{Pass it the name of the bucket to be deleted}
}
\value{
Returns 1 if successful
}
\description{
Delete a bucket that you have.
}
\details{
Make sure the name is correct before calling. As it is irreversable.
}

