% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_acl}
\alias{get_acl}
\title{Get ACL of bucket location}
\usage{
get_acl(bucketName, key, filename = 0L)
}
\arguments{
\item{bucketName}{The name of the bucket}

\item{key}{The location of the folder/file.}

\item{filename}{The location you want to store the file.}
}
\value{
Places ACL information into the file you chose and
returns a 1
}
\description{
Get ACL of bucket location
}
\details{
This function gets the ACL of a file and puts all
the information into a file.
}

