% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_logging}
\alias{get_logging}
\title{Get logging Settings of bucket}
\usage{
get_logging(bucketName, filename)
}
\arguments{
\item{bucketName}{The name of the bucket you want the settings}

\item{filename}{The location to place the logging settings into}
}
\value{
Returns a file with logging settings.
}
\description{
Get logging Settings of bucket
}
\details{
returns the settings for logging.
}

