% Generated by roxygen2 (4.0.1): do not edit by hand
\name{put_object}
\alias{put_object}
\title{Uploads an object to the S3 bucket specified.}
\usage{
put_object(bucketName, storage_location, filename, contentType = 0L)
}
\arguments{
\item{bucketName}{The name of the bucket you want to upload to.}

\item{storage_location}{The location you want to place the file.  For example
"file.newname" uploads it  with the name file.newname. Do not start with a "/"
unless you want a new blank folder.}

\item{filename}{The name of the file to be uploaded.  This is typically the location
on your computer.}
}
\value{
Returns 1 if successful.
}
\description{
Uploads an object to the S3 bucket specified.
}
\details{
Uploads a file to the S3 bucket you specify in the location specified.
}

