% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_connect}
\alias{S3_connect}
\title{Set the access_key and secret_key}
\usage{
S3_connect(access_key, secret_key, hostname = "s3.amazonaws.com")
}
\arguments{
\item{access_key}{Put your AWS access_key here to connect
to S3.}

\item{secret_key}{Put your AWS secret_key here to connect
to S3.}

\item{hostname}{Put a hostname such as 's3-eu-west-1.amazonaws.com'
to specify a name other than s3.amazonaws.com}
}
\value{
There is no return currently.
}
\description{
Set the access_key and secret_key
}
\details{
Before any other function use S3Connect to set
the access and secret key.
}

