% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_copy_object}
\alias{S3_copy_object}
\title{Copy an object from one bucket to another.}
\usage{
S3_copy_object(sourceBucketName, sourceKey, destinationBucketName,
  destinationKey)
}
\arguments{
\item{sourceBucketName}{The name of the bucket the file is in originally.}

\item{sourceKey}{The location the file is including folder as well as the
file name.}

\item{destinationBucketName}{The name of the bucket you want the file copied to.}

\item{destinationKey}{The location you want the file copied too.}
}
\value{
Returns 1 if successful
}
\description{
Copy an object from one bucket to another.
}
\details{
This function copies a file from one bucket to another
bucket and can place it within a new folder.
}

