% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_delete_bucket}
\alias{S3_delete_bucket}
\title{Delete a bucket that you have.}
\usage{
S3_delete_bucket(bucketName)
}
\arguments{
\item{bucketName}{Pass it the name of the bucket to be deleted}
}
\value{
Returns 1 if successful
}
\description{
Delete a bucket that you have.
}
\details{
Make sure the name is correct before calling. As it is irreversable.
}

