% Generated by roxygen2 (4.0.1): do not edit by hand
\name{S3_delete_object}
\alias{S3_delete_object}
\title{Delete's a file/object within the bucket}
\usage{
S3_delete_object(bucketName, key)
}
\arguments{
\item{bucketName}{The name of the bucket that contains the file/object}

\item{key}{The name of the file and location of it.  For example "filename"
will delete a file named filename in the root of the bucket.  "test/filename"
will delete one within the test folder.}
}
\value{
returns 1 if successful.
}
\description{
Delete's a file/object within the bucket
}
\details{
It will delete an object within the bucket.
}

