% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_get_acl}
\alias{S3_get_acl}
\title{Get ACL of bucket location}
\usage{
S3_get_acl(bucketName, key, filename = 0L)
}
\arguments{
\item{bucketName}{The name of the bucket}

\item{key}{The location of the folder/file.}

\item{filename}{The location you want to store the file.}
}
\value{
Places ACL information into the file you chose and
returns a 1
}
\description{
Get ACL of bucket location
}
\details{
This function gets the ACL of a file and puts all
the information into a file. View the AWS documentation on what ACL is \href{http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html}{here.}
}

