% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_get_logging}
\alias{S3_get_logging}
\title{Get logging Settings of bucket}
\usage{
S3_get_logging(bucketName, filename)
}
\arguments{
\item{bucketName}{The name of the bucket you want the settings}

\item{filename}{The location to place the logging settings into}
}
\value{
Returns a file with logging settings.
}
\description{
Get logging Settings of bucket
}
\details{
returns the settings for logging.
}

