% Generated by roxygen2 (4.0.1): do not edit by hand
\name{S3_get_object}
\alias{S3_get_object}
\title{Download a file from a bucket}
\usage{
S3_get_object(bucketName, key, filename = 0L)
}
\arguments{
\item{bucketName}{The name of the bucket where the file currently is.}

\item{key}{The location of the file for example "test/filename" is
a file located in the folder test.}

\item{filename}{The name of the location you want the file downloaded to.}
}
\value{
Returns 1 if successful.
}
\description{
Download a file from a bucket
}
\details{
This function downloads a file from the bucket that you specify.
}

