% Generated by roxygen2 (4.0.1): do not edit by hand
\name{S3_head_object}
\alias{S3_head_object}
\title{Gets the response properties of the object.}
\usage{
S3_head_object(bucketName, key)
}
\arguments{
\item{bucketName}{The name of the bucket where the file lies.}

\item{key}{The location of the file}
}
\value{
Returns the headers of the file.
}
\description{
Gets the response properties of the object.
}
\details{
This function lists all the properties of the object.
}

