% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_list_bucket}
\alias{S3_list_bucket}
\title{List all contents of the bucket}
\usage{
S3_list_bucket(bucketName, prefix = "", allDetails = 0L)
}
\arguments{
\item{bucketName}{The name of the bucket you want listed.}

\item{prefix}{If you want only a certain part of the bucket, for example
everything within "/blogs/test", then put that.}

\item{allDetails}{True/False if you want to see all the details or condensed version.}
}
\value{
Outputs in command line all info.
}
\description{
List all contents of the bucket
}
\details{
It will list all the contents of the bucket.  In the future it will be in a data frame.
}

