% Generated by roxygen2 (4.0.1): do not edit by hand
\name{S3_set_acl}
\alias{S3_set_acl}
\title{Set The ACL of bucket}
\usage{
S3_set_acl(bucketName, key, filename)
}
\arguments{
\item{bucketName}{The name of the bucket you want the settings}

\item{key}{The location of the folder/file}

\item{filename}{A file with ACL configs in}
}
\value{
Returns 1 if success
}
\description{
Set The ACL of bucket
}
\details{
returns the settings for logging.
}

