% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_test_bucket}
\alias{S3_test_bucket}
\title{Test the bucket and connection}
\usage{
S3_test_bucket(bucketName)
}
\arguments{
\item{bucketName}{Put a bucketName that exists that you use to test it.}
}
\value{
It will return all the headers of the call
}
\description{
Test the bucket and connection
}
\details{
This function allows you to test to see if everything is working correctly.
It tests that your two keys are correct.
}

