% Generated by roxygen2 (4.0.1): do not edit by hand
\name{S3Connect}
\alias{S3Connect}
\title{Set the access_key and secret_key}
\usage{
S3Connect(access_key, secret_key)
}
\arguments{
\item{access_key}{Put your AWS access_key here to connect
to S3.}

\item{secret_key}{Put your AWS secret_key here to connect
to S3.}
}
\value{
There is no return currently.
}
\description{
Set the access_key and secret_key
}
\details{
Before any other function use S3Connect to set
the access and secret key.
}

