% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_logging}
\alias{set_logging}
\title{Set the logging Settings for a bucket}
\usage{
set_logging(bucketName, targetBucket, filename)
}
\arguments{
\item{bucketName}{The name of the bucket to set the logging.}

\item{targetBucket}{The location of the bucket where you
want the logs to go.}

\item{filename}{The file with the logging settings inside}
}
\value{
returns a 1 if successful
}
\description{
Set the logging Settings for a bucket
}
\details{
Set the logging settings for a bucket.
}

