% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_bucket}
\alias{test_bucket}
\title{Test the bucket and connection}
\usage{
test_bucket(bucketName)
}
\arguments{
\item{bucketName}{Put a bucketName that exists that you use to test it.}
}
\value{
It will return all the headers of the call
}
\description{
Test the bucket and connection
}
\details{
This function allows you to test to see if everything is working correctly.
It tests that your two keys are correct.
}

