% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S3_set_acl}
\alias{S3_set_acl}
\title{Set The ACL of bucket}
\usage{
S3_set_acl(bucketName, key, filename)
}
\arguments{
\item{bucketName}{The name of the bucket you want the settings}

\item{key}{The location of the folder/file}

\item{filename}{A file with ACL configs in}
}
\value{
Returns 1 if success
}
\description{
Set The ACL of bucket
}
\details{
returns the settings for logging. View the AWS documentation on what ACL is \href{http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html}{here.}
}

